refactor : delete 할 알림이 하나도 없으면 쿼리 날리지 않고 early return
#345

+++ backend/emm-sale/src/main/java/com/emmsale/notification/application/UpdateNotificationCommandService.java
@@ -45,6 +45,14 @@ public class UpdateNotificationCommandService {
         .map(UpdateNotification::getId)
         .collect(Collectors.toList());
 
+    if (hasNoNotificationToDeleteBy(deleteIds)) {
+      return;
+    }
+
     updateNotificationRepository.deleteAllByIdInBatch(deletedIds);
+  }
+
+  private static boolean hasNoNotificationToDeleteBy(final List<Long> deleteIds) {
+    return deleteIds.isEmpty();
   }
 }
